refactor: 이미지 관련 패키지 구분
